+++ libdispatch/dfile.c
@@ -672,6 +672,8 @@ Open a netCDF file with the contents taken from a block of memory.
 
 \param path Must be non-null, but otherwise only used to set the dataset name.
 
+\param mode the mode flags; Note that this procedure uses a limited set of flags because it forcibly sets NC_NOWRITE|NC_DISKLESS|NC_INMEMORY.
+
 \param size The length of the block of memory being passed.
  
 \param memory Pointer to the block of memory containing the contents
@@ -684,7 +686,31 @@ stored.
 
 \returns ::NC_ENOMEM Out of memory.
 
-\returns Various other netcdf classic and netcdf 4 errors.
+\returns ::NC_EDISKLESS diskless io is not enabled for fails.
+
+\returns ::NC_EINVAL, etc. other errors also returned by nc_open.
+
+<h1>Examples</h1>
+
+Here is an example using nc_open_mem() to open an existing netCDF dataset
+named foo.nc for read-only, non-shared access. It differs from the nc_open()
+example in that it assumes the contents of foo.nc have been read into memory.
+
+@code
+#include <netcdf.h>
+#include <netcdf_mem.h>
+   ... 
+int status = NC_NOERR;
+int ncid;
+size_t size;
+void* memory;
+   ... 
+size = <compute file size of foo.nc in bytes>;
+memory = malloc(size);
+   ... 
+status = nc_open_mem("foo.nc", 0, size, memory, &ncid);
+if (status != NC_NOERR) handle_error(status);
+@endcode
 */
 int
 nc_open_mem(const char* path, int mode, size_t size, void* memory, int* ncidp)